+++ pymtl3/dsl/ComponentLevel1.py
@@ -77,6 +77,9 @@ class ComponentLevel1( NamedObject ):
     s._dsl.upblks.add( blk )
     return blk
 
+  def get_update_block( s, name ):
+    return s._dsl.name_upblk[ name ]
+
   def add_constraints( s, *args ):
     if isinstance( s, Placeholder ):
       raise InvalidPlaceholderError( "Cannot define constraints "
@@ -93,7 +96,7 @@ class ComponentLevel1( NamedObject ):
                               " is not implemented in {}".format( s.__class__.__name__ ) )
 
   #-----------------------------------------------------------------------
-  # elaborate
+  # elaborate-related public APIs
   #-----------------------------------------------------------------------
 
   # We refactor the monolithic elaborate function into smaller functions
@@ -122,7 +125,7 @@ class ComponentLevel1( NamedObject ):
     s._elaborate_declare_vars()
     s._elaborate_collect_all_vars()
   #-----------------------------------------------------------------------
-  # Public APIs (only can be called after elaboration)
+  # Post-elaborate public APIs (can only be called after elaboration)
   #-----------------------------------------------------------------------
 
   def is_component( s ):
@@ -134,130 +137,3 @@ class ComponentLevel1( NamedObject ):
   def is_interface( s ):
     return False
 
-  def get_update_block( s, name ):
-    return s._dsl.name_upblk[ name ]
-
-  def get_update_block_host_component( s, blk ):
-    try:
-      assert s._dsl.elaborate_top is s, "Getting update block host component " \
-                                    "is only allowed at top, but this API call " \
-                                    "is on {}.".format( "top."+repr(s)[2:] )
-      return s._dsl.all_upblk_hostobj[ blk ]
-    except AttributeError:
-      raise NotElaboratedError()
-
-  def get_update_blocks( s ):
-    return s._dsl.upblks
-
-  def get_all_update_blocks( s ):
-    try:
-      assert s._dsl.elaborate_top is s, "Getting all update blocks " \
-                                  "is only allowed at top, but this API call " \
-                                  "is on {}.".format( "top."+repr(s)[2:] )
-      return s._dsl.all_upblks
-    except AttributeError:
-      raise NotElaboratedError()
-
-  def get_component_level( s ):
-    try:
-      return s._dsl.level
-    except AttributeError:
-      raise NotElaboratedError()
-
-  def get_all_explicit_constraints( s ):
-    try:
-      assert s._dsl.elaborate_top is s, "Getting all explicit constraints " \
-                                    "is only allowed at top, but this API call " \
-                                    "is on {}.".format( "top."+repr(s)[2:] )
-      return s._dsl.all_U_U_constraints
-    except AttributeError:
-      raise NotElaboratedError()
-
-  def get_child_components( s ):
-    assert s._dsl.constructed
-    ret = set()
-    stack = []
-    for (name, obj) in s.__dict__.iteritems():
-      if   isinstance( name, basestring ): # python2 specific
-        if not name.startswith("_"): # filter private variables
-          stack.append( obj )
-    while stack:
-      u = stack.pop()
-      if   isinstance( u, ComponentLevel1 ):
-        ret.add( u )
-      # ONLY LIST IS SUPPORTED
-      elif isinstance( u, list ):
-        stack.extend( u )
-    return ret
-
-  def get_all_components( s ):
-    try:
-      return s._dsl.all_components
-    except AttributeError:
-      return s._collect_all( lambda x: isinstance( x, ComponentLevel1 ) )
-
-  def get_all_object_filter( s, filt ):
-    assert callable( filt )
-    try:
-      return set( [ x for x in s._dsl.all_components if filt(x) ] )
-    except AttributeError:
-      return s._collect_all( filt )
-
-  def delete_component_by_name( s, name ):
-
-    # This nested delete function is to create an extra layer to properly
-    # call garbage collector. If we can make sure it is collected
-    # automatically and fast enough, we might remove this extra layer
-    #
-    # EDIT: After experimented w/ and w/o gc.collect(), it seems like it
-    # is eventually collected, but sometimes the intermediate memory
-    # footprint might reach up to gigabytes, so let's keep the
-    # gc.collect() for now
-
-    def _delete_component_by_name( parent, name ):
-      obj = getattr( parent, name )
-      top = s._dsl.elaborate_top
-
-      # Remove all components and uncollect metadata
-
-      removed_components = obj.get_all_components()
-      top._dsl.all_components -= removed_components
-
-      for x in removed_components:
-        assert x._dsl.elaborate_top is top
-        top._uncollect_vars( x )
-
-      for x in obj._collect_all():
-        del x._dsl.parent_obj
-
-      delattr( s, name )
-
-    _delete_component_by_name( s, name )
-    import gc
-    gc.collect()
-
-  def add_component_by_name( s, name, obj ):
-    assert not hasattr( s, name )
-    NamedObject.__setattr__ = NamedObject.__setattr_for_elaborate__
-    setattr( s, name, obj )
-    del NamedObject.__setattr__
-
-    top = s._dsl.elaborate_top
-
-    added_components = obj.get_all_components()
-    top._dsl.all_components |= added_components
-
-    for c in added_components:
-      c._dsl.elaborate_top = top
-      top._collect_vars( c )
-
-  def apply( s, *args ):
-
-    if isinstance(args[0], list):
-      assert len(args) == 1
-      for step in args[0]:
-        step( s )
-
-    elif len(args) == 1:
-      assert callable( args[0] )
-      args[0]( s )
